Fix wrong bank journal when using FACTURE_PAYMENTS_ON_DIFFERENT_THIRDPARTIES_BILLS
#32393
When we use this optiion "FACTURE_PAYMENTS_ON_DIFFERENT_THIRDPARTIES_BILLS" if the same payment groups the invoices of several subsidiaries (same parent company) thanks to this option, the transfer of the bank journal becomes crazy! The amount proposed for the bank's accounting account is equal to the amount paid multiplied by the number of subsidiaries included. The amount proposed for the subsidiaries' accounting account is equal to the amount paid multiplied by the number of subsidiaries included and by the number of different invoices for each subsidiary.

This patch is very partial, it only allows to quickly unblock the blockage observed by proposing the transfer to accounting of the correct amounts for the bank account and the subsidiary having carried out the monetary flow.

A resumption of this option is necessary to make it usable in compliance with the accounting rules!